foosball-frontend-6 Add a player
Implemented the ability to add a new player

+++ src/app/players/player/player.component.ts
@@ -1,27 +1,32 @@
-import { Component, Input, OnDestroy } from '@angular/core';
+import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { Player } from "./player.model";
 import { PlayersService } from "../players.service";
 import { FormControl, FormGroup, Validators } from "@angular/forms";
 import { Subscription } from "rxjs";
 import { ObjectToFormGroup } from "../../types/form-group.type";
 
+// ToDo Adding spinner and waiting for the save on the backend to resolve
+
 @Component({
-    selector: 'app-player[player]',
+    selector: 'app-player',
     templateUrl: './player.component.html',
     styleUrls: ['./player.component.scss']
 })
-export class PlayerComponent implements OnDestroy {
-    modifiedAttributes = new Set<keyof Player>();
-    playerForm!: FormGroup<ObjectToFormGroup<Omit<Player, 'id'>>>
+export class PlayerComponent implements OnInit, OnDestroy {
+    @Output() destroy = new EventEmitter<void>();
+
+    protected modifiedAttributes = new Set<keyof Player>();
+    protected playerForm!: FormGroup<ObjectToFormGroup<Omit<Player, 'id'>>>
+    protected newPlayer: boolean = false;
+    private _player?: Player;
+
     private formSubscription?: Subscription;
 
     constructor(private playersService: PlayersService) {
     }
 
-    private _player!: Player;
-
     get player(): Player {
-        return this._player;
+        return this._player!
     }
 
     /**
@@ -36,6 +41,17 @@ export class PlayerComponent implements OnDestroy {
         this.formSubscription = this.onPlayerFormChange();
     }
 
+    ngOnInit() {
+        if (!this._player) {
+            this.newPlayer = true;
+            this.player = {
+                id: -1,
+                firstName: '',
+                lastName: ''
+            }
+        }
+    }
+
     ngOnDestroy() {
         this.formSubscription?.unsubscribe();
     }
@@ -44,26 +60,39 @@ export class PlayerComponent implements OnDestroy {
      * Cancel the editing and resets to the previous values.
      */
     onCancelEdit() {
-        this.playerForm.reset()
+        if (!this.newPlayer) {
+            this.playerForm.reset()
+        } else {
+            this.destroy.emit()
+        }
     }
 
     /**
      * Saves the changes to the database.
      */
     onSaveEdit() {
-        this.playerForm.markAsTouched();
-        if (this.playerForm.invalid) return;
+        this.playerForm.markAllAsTouched();
+        if (this.playerForm.invalid || !this._player) return;
 
-        this.playersService.editPlayer(this._player.id, {
+        const updatedPlayer = {
             ...this._player,
             ...this.playerForm.value
-        })
+        };
+
+        if (!this.newPlayer) {
+            this.playersService.editPlayer(this._player.id, updatedPlayer)
+        } else {
+            this.playersService.addPlayer(updatedPlayer)
+            this.destroy.emit()
+        }
     }
 
     /**
      * Deletes the player from the database.
      */
     onDelete() {
+        if (!this._player || this.newPlayer) return;
+
         this.playersService.deletePlayer(this._player.id)
     }
 
@@ -92,7 +121,7 @@ export class PlayerComponent implements OnDestroy {
             this.modifiedAttributes.clear();
             Object.entries(modifiedPlayer).forEach(([key, value]) => {
                 const playerAttribute = key as keyof Player
-                if (value !== this._player[playerAttribute]) {
+                if (value !== this._player![playerAttribute]) {
                     this.modifiedAttributes.add(playerAttribute)
                 }
             })
